+++ frontend/src/components/meeting/MeetingMap.js
@@ -53,12 +53,13 @@ function MeetingMap(props) {
           center={clickPosition || currentLocation}
           style={{ width: '100%', height: '500px' }}
           level={3}
-          onClick={(_t, mouseEvent) =>
+          onClick={(_t, mouseEvent) => {
             setClickPosition({
               lat: mouseEvent.latLng.getLat(),
               lng: mouseEvent.latLng.getLng(),
-            })
-          }
+            });
+            setSelectedMarker(null);
+          }}
           onCreate={setMap}
         >
           <MarkerClusterer averageCenter={true} minLevel={10}>
@@ -73,6 +74,7 @@ function MeetingMap(props) {
                     },
                   });
                   setInfo(null);
+                  setDescription('');
                 }}
               />
             )}
@@ -84,6 +86,7 @@ function MeetingMap(props) {
                   onClick={() => {
                     setInfo(marker);
                     setSelectedMarker(marker);
+                    setDescription('');
                   }}
                 >
                   {info && info.content === marker.content && (
@@ -124,7 +127,7 @@ function MeetingMap(props) {
             <p>
               Add description about this place!{'  '}
               <Input
-                defaultValue={description}
+                value={description}
                 size="mini"
                 onChange={(e) => setDescription(e.target.value)}
               />
